use a single wrapper file (#124)
