+++ R/RcppExports.R
@@ -1,6 +1,10 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+pseudo_obs_cpp <- function(x, ties_method) {
+    .Call(`_rvinecopulib_pseudo_obs_cpp`, x, ties_method)
+}
+
 bicop_check_cpp <- function(bicop_r) {
     invisible(.Call(`_rvinecopulib_bicop_check_cpp`, bicop_r))
 }
@@ -43,10 +47,6 @@ bicop_par_to_tau_cpp <- function(bicop_r) {
 
 bicop_tau_to_par_cpp <- function(bicop_r, tau) {
     .Call(`_rvinecopulib_bicop_tau_to_par_cpp`, bicop_r, tau)
-}
-
-pseudo_obs_cpp <- function(x, ties_method) {
-    .Call(`_rvinecopulib_pseudo_obs_cpp`, x, ties_method)
 }
 
 rvine_matrix_check_cpp <- function(matrix) {
